split server logic from port logic
