+++ api/package.json
@@ -63,12 +63,14 @@
     "dotenv": "^6.1.1",
     "jest": "^24.8.0",
     "mocha": "^6.1.4",
+    "reflect-metadata": "^0.1.13",
     "supertest": "^4.0.2",
     "ts-jest": "^24.0.2",
     "ts-loader": "^6.0.0",
     "ts-node": "^8.1.0",
     "tslint": "^5.8.0",
     "tsoa": "^2.3.81",
+    "typeorm": "^0.2.18",
     "typescript": "^3.4.5",
     "webpack": "^4.31.0",
     "webpack-cli": "^3.3.2"
