+++ api/src/server/index.ts
@@ -1,31 +1,24 @@
-import * as express from 'express';
-import * as dotenv from 'dotenv';
-import * as bodyParser from 'body-parser';
-import * as swaggerUI from 'swagger-ui-express';
+import express, { Express } from 'express';
+import bodyParser from 'body-parser';
+import swaggerUI from 'swagger-ui-express';
 
 import { RegisterRoutes } from './routes/routes';
-import * as swaggerJSON from './swagger/swagger.json';
+import swaggerJSON from './swagger/swagger.json';
 
-if (process.env.NODE_ENV === 'development') {
-  dotenv.config();
+function addSwagger(app: Express) {
+  app.use('/swagger.json', express.static(__dirname + './swagger/swagger.json'));
+  app.use('/api-docs', swaggerUI.serve, swaggerUI.setup(swaggerJSON));
 }
 
-const app = express();
+function addBodyParser(app: Express) {
+  app.use(bodyParser.json());
+  app.use(bodyParser.urlencoded({ extended: true }));
+}
 
-app.set('port', process.env.PORT || 3000);
+function init(app: Express) {
+  addSwagger(app);
+  addBodyParser(app);
+  RegisterRoutes(app);
+}
 
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-
-app.use('/swagger.json', express.static(__dirname + './swagger/swagger.json'));
-app.use('/api-docs', swaggerUI.serve, swaggerUI.setup(swaggerJSON));
-
-RegisterRoutes(app);
-
-app.listen(app.get('port'), () => {
-  console.log(('App is running at http://localhost:%d in %s mode'),
-    app.get('port'), app.get('env'));
-  console.log('Press CTRL-C to stop\n');
-});
-
-export default app;
+export default init;
