+++ api/tsconfig.json
@@ -13,6 +13,7 @@
     "rootDir": "src",
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
+    "esModuleInterop": true,
     "sourceMap": true,
     "baseUrl": ".",
     "moduleResolution": "node",
